backend/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0

# Validation
pydantic==2.10.4

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.18

# Migrations
alembic==1.14.0

# Development
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
pytest-cov==6.0.0
